qa fix invalid delete

+++ app.py
@@ -18,17 +18,19 @@ def home():
 @app.route("/add", methods=['POST'])
 def add():
     task_to_add = request.form['add']
-    dao.connect_to_db()
-    dao.insert_without_id(task_to_add)
-    return redirect(request.referrer)
+    if task_to_add is not "":
+        dao.connect_to_db()
+        dao.insert_without_id(task_to_add)
+        return redirect(request.referrer)
 
 
 @app.route("/delete", methods=['POST'])
 def delete():
     id_to_delete = request.form['delete']
-    dao.connect_to_db()
-    dao.delete_by_id(id_to_delete)
-    return redirect(request.referrer)
+    if id_to_delete is not "":
+        dao.connect_to_db()
+        dao.delete_by_id(id_to_delete)
+        return redirect(request.referrer)
 
 
 if __name__ == "__main__":
